requirements.txt
mysql.connector
pandas
sqlalchemy
pymysql
